% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Deconstruct data into multiple values}
\usage{
vars \%<-\% data
}
\arguments{
\item{vars}{[R names] Valid R names separated by colon.}

\item{data}{[object] R object of the same length as the number of \code{vars}
names.}
}
\value{
Nothing. Assigns LHS names to current scope.
}
\description{
\code{consign} pipe allows destructing data into multiple variables at once.
Internally \code{rlang::set_names} is used to name the RHS elements to current
scope.
}
\examples{
val1 : val_2 : someval \%<-\% c(2, 4, 9)
c(val1, val_2, someval)

x:y:z \%<-\% list(c(4,5,12), TRUE, list(a=5,b=9))
x; y; z

}
\seealso{
Other result assemblers: 
\code{\link{\%$>\%}()},
\code{\link{\%->\%}()},
\code{\link{\%to\%}()},
\code{\link{conserve}()},
\code{\link{control}()}
}
\concept{result assemblers}
