% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_pretty}
\alias{str_pretty}
\title{Prettier Printing of object overview}
\usage{
str_pretty(x, ...)
}
\arguments{
\item{x}{[R \code{object}] Any object.}

\item{...}{[named arguments] Extra arguments to pass to \code{str}.}
}
\value{
Nothing. Only side effect being output to stdout.
}
\description{
\code{str_pretty} is a wrapper around \code{str} that tries to use better, more
distinct identifiers in the object overview.
}
\examples{
str_pretty(lm(Sepal.Length ~ ., data = iris))

nst_lst <- list()
nst_lst$a <- list(c(9,2,4, 12), g=c(2,4,9,7), h=list(m=4,n=7,d=12))
nst_lst$b <- c(4, 3, 9, 10, 3, 16, 1, 7)
nst_lst$f <- c("a", "b")
nst_lst$c <- 4
nst_lst$d <- c(7,9)
str_pretty(nst_lst)

# Pass `str` arguments to modify behavior
str_pretty(lm(Sepal.Length ~ ., data = iris), give.attr = TRUE)

}
